add separate faq styling/text-- to be changed with the updated copy
