+++ site/src/components.d.ts
@@ -43,6 +43,18 @@ export namespace Components {
          */
         "pageTitle"?: string;
     }
+    interface PageFaqApplicationStatus {
+        /**
+          * A list of entries to display in the FAQ see: FaqData.ts see: app-root.tsx
+         */
+        "data"?: { sectionTitle: string, questions: { question: string, answer: () => string }[] }[];
+    }
+    interface PageFaqEligibility {
+        /**
+          * A list of entries to display in the FAQ see: FaqData.ts see: app-root.tsx
+         */
+        "data"?: { sectionTitle: string, questions: { question: string, answer: () => string }[] }[];
+    }
     interface PageFaqEs {
         /**
           * A list of entries to display in the FAQ see: FaqData.ts see: app-root.tsx
@@ -52,6 +64,12 @@ export namespace Components {
           * The page's title
          */
         "pageTitle"?: string;
+    }
+    interface PageFaqPollWorker {
+        /**
+          * A list of entries to display in the FAQ see: FaqData.ts see: app-root.tsx
+         */
+        "data"?: { sectionTitle: string, questions: { question: string, answer: () => string }[] }[];
     }
     interface PageForm {
         /**
@@ -191,11 +209,29 @@ declare global {
         prototype: HTMLPageFaqElement;
         new (): HTMLPageFaqElement;
     };
+    interface HTMLPageFaqApplicationStatusElement extends Components.PageFaqApplicationStatus, HTMLStencilElement {
+    }
+    var HTMLPageFaqApplicationStatusElement: {
+        prototype: HTMLPageFaqApplicationStatusElement;
+        new (): HTMLPageFaqApplicationStatusElement;
+    };
+    interface HTMLPageFaqEligibilityElement extends Components.PageFaqEligibility, HTMLStencilElement {
+    }
+    var HTMLPageFaqEligibilityElement: {
+        prototype: HTMLPageFaqEligibilityElement;
+        new (): HTMLPageFaqEligibilityElement;
+    };
     interface HTMLPageFaqEsElement extends Components.PageFaqEs, HTMLStencilElement {
     }
     var HTMLPageFaqEsElement: {
         prototype: HTMLPageFaqEsElement;
         new (): HTMLPageFaqEsElement;
+    };
+    interface HTMLPageFaqPollWorkerElement extends Components.PageFaqPollWorker, HTMLStencilElement {
+    }
+    var HTMLPageFaqPollWorkerElement: {
+        prototype: HTMLPageFaqPollWorkerElement;
+        new (): HTMLPageFaqPollWorkerElement;
     };
     interface HTMLPageFormElement extends Components.PageForm, HTMLStencilElement {
     }
@@ -296,7 +332,10 @@ declare global {
         "page-contact": HTMLPageContactElement;
         "page-election-officials": HTMLPageElectionOfficialsElement;
         "page-faq": HTMLPageFaqElement;
+        "page-faq-application-status": HTMLPageFaqApplicationStatusElement;
+        "page-faq-eligibility": HTMLPageFaqEligibilityElement;
         "page-faq-es": HTMLPageFaqEsElement;
+        "page-faq-poll-worker": HTMLPageFaqPollWorkerElement;
         "page-form": HTMLPageFormElement;
         "page-info": HTMLPageInfoElement;
         "page-jurisdiction": HTMLPageJurisdictionElement;
@@ -351,6 +390,18 @@ declare namespace LocalJSX {
          */
         "pageTitle"?: string;
     }
+    interface PageFaqApplicationStatus {
+        /**
+          * A list of entries to display in the FAQ see: FaqData.ts see: app-root.tsx
+         */
+        "data"?: { sectionTitle: string, questions: { question: string, answer: () => string }[] }[];
+    }
+    interface PageFaqEligibility {
+        /**
+          * A list of entries to display in the FAQ see: FaqData.ts see: app-root.tsx
+         */
+        "data"?: { sectionTitle: string, questions: { question: string, answer: () => string }[] }[];
+    }
     interface PageFaqEs {
         /**
           * A list of entries to display in the FAQ see: FaqData.ts see: app-root.tsx
@@ -360,6 +411,12 @@ declare namespace LocalJSX {
           * The page's title
          */
         "pageTitle"?: string;
+    }
+    interface PageFaqPollWorker {
+        /**
+          * A list of entries to display in the FAQ see: FaqData.ts see: app-root.tsx
+         */
+        "data"?: { sectionTitle: string, questions: { question: string, answer: () => string }[] }[];
     }
     interface PageForm {
         /**
@@ -459,7 +516,10 @@ declare namespace LocalJSX {
         "page-contact": PageContact;
         "page-election-officials": PageElectionOfficials;
         "page-faq": PageFaq;
+        "page-faq-application-status": PageFaqApplicationStatus;
+        "page-faq-eligibility": PageFaqEligibility;
         "page-faq-es": PageFaqEs;
+        "page-faq-poll-worker": PageFaqPollWorker;
         "page-form": PageForm;
         "page-info": PageInfo;
         "page-jurisdiction": PageJurisdiction;
@@ -489,7 +549,10 @@ declare module "@stencil/core" {
             "page-contact": LocalJSX.PageContact & JSXBase.HTMLAttributes<HTMLPageContactElement>;
             "page-election-officials": LocalJSX.PageElectionOfficials & JSXBase.HTMLAttributes<HTMLPageElectionOfficialsElement>;
             "page-faq": LocalJSX.PageFaq & JSXBase.HTMLAttributes<HTMLPageFaqElement>;
+            "page-faq-application-status": LocalJSX.PageFaqApplicationStatus & JSXBase.HTMLAttributes<HTMLPageFaqApplicationStatusElement>;
+            "page-faq-eligibility": LocalJSX.PageFaqEligibility & JSXBase.HTMLAttributes<HTMLPageFaqEligibilityElement>;
             "page-faq-es": LocalJSX.PageFaqEs & JSXBase.HTMLAttributes<HTMLPageFaqEsElement>;
+            "page-faq-poll-worker": LocalJSX.PageFaqPollWorker & JSXBase.HTMLAttributes<HTMLPageFaqPollWorkerElement>;
             "page-form": LocalJSX.PageForm & JSXBase.HTMLAttributes<HTMLPageFormElement>;
             "page-info": LocalJSX.PageInfo & JSXBase.HTMLAttributes<HTMLPageInfoElement>;
             "page-jurisdiction": LocalJSX.PageJurisdiction & JSXBase.HTMLAttributes<HTMLPageJurisdictionElement>;
