+++ site/src/components/app-root/app-root.tsx
@@ -1,6 +1,6 @@
 import { Component, FunctionalComponent, h, State } from "@stencil/core";
 
-import { FaqData, FaqESData, PartnerList, Social } from "../../data";
+import { FaqApplicationData, FaqEligibilityData, FaqPollWorkerData, FaqESData, PartnerList, Social } from "../../data";
 import { Partner } from "../../data/types";
 import { parseQueryString, shuffle } from "../../util";
 
@@ -22,7 +22,7 @@ const Nav: FunctionalComponent<NavProps> = ( { onSelectNavItem, formPath, partne
          <li>
             <stencil-route-link
                url="/faq"
-               urlMatch={["/faq", "/faq-es", "/contact", "/faq-contact"]}
+               urlMatch={["/faq", "/faq-es", "/faq-contact"]}
                onClick={onSelectNavItem}
             >FAQ</stencil-route-link>
          </li>
@@ -41,7 +41,7 @@ const Nav: FunctionalComponent<NavProps> = ( { onSelectNavItem, formPath, partne
          </li>
          <li>
             <stencil-route-link
-               url="/faq"
+               url="/contact"
                urlMatch={["/contact"]}
                onClick={onSelectNavItem}
             >Contact</stencil-route-link>
@@ -102,8 +102,14 @@ export class AppRoot {
             url: "/faq",
             component: "page-faq",
             componentProps: {
-               data: FaqData,
                pageTitle: "Frequently Asked Questions about Poll Working",
+            },
+         },
+         {
+            url: "/contact",
+            component: "page-contact",
+            componentProps: {
+               pageTitle: "Contact Us",
             },
          },
          {
@@ -156,18 +162,31 @@ export class AppRoot {
             url: "/testimonial",
             component: "page-testimonial",
          },
+         {
+            url: "/faq-eligibility",
+            component: "page-faq-eligibility",
+            componentProps: {
+               data: FaqEligibilityData,
+            },
+         },
+         {
+            url: "/faq-application-status",
+            component: "page-faq-application-status",
+            componentProps: {
+               data: FaqApplicationData,
+            },
+         },
+         {
+            url: "/faq-poll-worker",
+            component: "page-faq-poll-worker",
+            componentProps: {
+               data: FaqPollWorkerData,
+            },
+         },
 
          /*
           * legacy routes
           */
-         {
-            url: "/contact",
-            component: "page-faq",
-            componentProps: {
-               data: FaqData,
-               pageTitle: "Frequently Asked Questions about Poll Working",
-            },
-         },
          {
             url: "/redirector",
             component: "page-info",
