+++ site/src/components/page-contact/page-contact.tsx
@@ -5,7 +5,7 @@ import {Component, h, Host, Prop, State} from "@stencil/core";
     styleUrl: "page-contact.scss",
     shadow: false,
 })
-export class PageFaq {
+export class PageContact{
 
     /**
      * A list of entries to display in the FAQ
@@ -28,14 +28,32 @@ export class PageFaq {
         return (
             <Host>
                 <h1>{this.pageTitle || ""}</h1>
+                <h2> Press </h2>
                 <p>
                     For press inquiries, please contact <a
                     href="mailto:press@powerthepolls.org">press@powerthepolls.org</a>.
                 </p>
+                <h2> Partnership </h2>
                 <p>
                     To inquire about partnering with us, please contact <a
                     href="mailto:partners@powerthepolls.org">partners@powerthepolls.org</a>
                 </p>
+
+                <h2> Questions </h2>
+                <p>
+                    Frequently asked questions can be found <a
+                    href="https://www.powerthepolls.org/faq">here</a>, and resources can be found <a
+                    href="https://www.powerthepolls.org/resources">here</a>.
+                </p>
+
+
+                <h2> Other needs </h2>
+
+                <button class="cta" onClick={() => this.isModalOpen = true}>contact us</button>
+                <contact-modal
+                    isOpen={isModalOpen}
+                    onClose={() => this.isModalOpen = false}
+                />
             </Host>
         );
     }
