+++ site/src/components/page-election-officials/page-faq-application-status.scss
@@ -1,6 +1,6 @@
-page-contact {
+page-faq-application-status {
 
-   button.cta {
+   .back-button {
       border: 0;
       background-color: $color-font-highlight-contrast;
       padding: 0.8em 1em;
