+++ site/src/components/page-faq/page-faq.scss
@@ -50,85 +50,11 @@ page-faq {
       }
    }
 
-   .videos-container {
-      margin: 1em;
-      display: grid;
-      align-items: flex-start;
-      align-content: space-between;
-      justify-content: center;
-      column-gap: 2em;
-      grid-template-columns: 50% 50%;
-      @include on-mobile() {
-         grid-template-columns: 100%;
-         margin: 1em 0;
-      }
-
-      & > div {
-         justify-self: stretch;
-      }
-   }
-
-   .placeholder {
-      display: grid;
-      position: relative;
-      cursor: default;
-
-      .box {
-         width: 100%;
-         height: 100%;
-         padding: 0;
-         padding-bottom: 55%;
-         background-color: #000;
-      }
-
-      .text {
-         position: absolute;
-         color: $color-accent;
-         font-size: 1.1rem;
-         text-transform: uppercase;
-         text-align: center;
-         vertical-align: middle;
-         align-self: center;
-         justify-self: center;
-      }
-   }
-
-   video {
-      width: 100%;
-      height: 100%;
-      cursor: pointer;
-   }
-
-   .additional-note {
-      display: grid;
-      grid-template-columns: 30% 70%;
-      column-gap: 1.5em;
-      justify-content: flex-start;
-      margin: 2em 0;
-      @include on-mobile() {
-         display: flex;
-         flex-direction: column;
-         & > img {
-            margin-bottom: 1em !important;
-            max-height: 120px;
-         }
-      }
-      & > img {
-         margin: auto;
-         &.background {
-            padding: 8px;
-            background: white;
-         }
-      }
-      & > p {
-         margin: 0;
-      }
-   }
-
    .column {
      float: left;
      width: 33%;
      margin: 0 -5px;
+     position: relative;
    }
 
    .card {
@@ -140,8 +66,8 @@ page-faq {
       position: relative;
 
       p {
-         padding: .5em;
-         font-size: 10px;
+         padding: .8em;
+         font-size: 12px;
          position: absolute;
          bottom: 0;
          color: #1C1834;
@@ -167,8 +93,8 @@ page-faq {
    }
 
    img {
-      width:100px;
-      height:100px;
+      width:110px;
+      height:110px;
       object-fit:scale-down;
       display: block;
       margin-left: auto;
