+++ site/src/data/index.ts
@@ -1,5 +1,7 @@
 export { default as FaqESData } from "./FaqESData";
-export { default as FaqData } from "./FaqData";
+export { default as FaqApplicationData } from "./FaqApplicationData";
+export { default as FaqEligibilityData } from "./FaqEligibilityData";
+export { default as FaqPollWorkerData } from "./FaqPollWorkerData";
 export { default as Social } from "./Social";
 export { default as PartnerList } from "./PartnerList";
 export { default as ResourceChecklistText } from "./ResourceChecklistText";
